mongo_alert_consumer/requirements.txt
kafka-python-ng==2.2.2
pymongo==4.6.1
